+++ lib/pages/settings_icon/icon.dart
@@ -12,51 +12,6 @@ void debug(String x) {
   print(x);
 }
 
-void main(){
-  runApp(MaterialApp(home: Scaffold(body: IconDisplayer())));
-}
-
-class IconDisplayer extends StatefulWidget {
-  @override
-  _IconDisplayerState createState() => _IconDisplayerState();
-}
-
-class _IconDisplayerState extends State<IconDisplayer>{
-  double _time = 0;
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        body: Column(
-          children: [
-            Container(
-              decoration: BoxDecoration(
-                border: Border.all(color: Colors.pink, width: 2),
-              ),
-              width: 600,
-              height: 600,
-              child: SettingsIcon(_time),
-            ),
-            Slider(
-              min: 0,
-              max: 100,
-              value: _time * 100,
-              divisions: 100,
-              onChanged: (double value) {
-                setState( () {
-                  debug('value changed to $value');
-                  _time = value / 100;
-                });
-              }
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
 class SettingsIcon extends StatelessWidget {
   const SettingsIcon(this.time);
